requirements.txt
# update setuptools if python version is 3.12 or greater to get distutils
# - distutils is dropped from the python standard lib from 3.12.
# - certain packages (which?) require distutils
setuptools>=68.2.2; python_version >="3.12"
astropy>=5.1.1
chardet>=4.0.0
pandas==1.5.3

# for now: prevent installing tensorflow >= 2.16.0,
# which requires installing tf-keras separately
tensorflow[and-cuda]>=2.14.0,<2.16.0;platform_system == "Linux" and platform_machine=="x86_64"

# Linux ARM64 (aarch64)
tensorflow==2.14.0;platform_system == "Linux" and platform_machine=="aarch64"

# macos Intel
tensorflow==2.14.0;platform_system == "Darwin" and platform_machine=="x86_64"
# macos Apple Silicon
tensorflow-macos==2.12.0;platform_system == "Darwin" and platform_machine=="arm64"

# system specific tensorflow install, using environment markers (https://peps.python.org/pep-0508/#environment-markers)
gpflow==2.6.3; platform_system == "Darwin" and platform_machine=="arm64"
# issue with using: not (environment_marker)
gpflow>=2.9.0; platform_system != "Darwin" or platform_machine!="arm64"


gpytorch==1.10
matplotlib>=3.6.2
scipy>=1.9.3
tables>=3.7.0
netCDF4==1.6.2
xarray>=2022.10.0
gast>=0.4.0
pyproj>=3.4.0
seaborn>=0.11.2
jupyter==1.0.0
scikit-learn==1.2.2
numba>=0.56.4
pytest>=7.2.0
dataclasses-json==0.5.7
global-land-mask==1.0.0
cartopy==0.22.0
fastparquet>=2024.2.0
pyarrow>=15.0.2



# for generating documentation - requires pandoc to be installed to generate (on linux: sudo apt install pandoc)
Sphinx>=5.0.2
nbsphinx>=0.9.3
numpydoc>=1.6.0
sphinxemoji>=0.2.0
sphinx-rtd-theme>=1.3.0
